Expose Halide infrastructure
