+++ tests/halide.rs
@@ -1,402 +1,3 @@
-use num::{ToPrimitive, Zero};
-use ruler::*;
-use z3::ast::Ast;
-
-type Constant = i64;
-
-egg::define_language! {
-  pub enum Pred {
-    Lit(Constant),
-    "abs" = Abs(Id),
-    "<" = Lt([Id;2]),
-    "<=" = Leq([Id;2]),
-    "==" = Eq([Id;2]),
-    "!=" = Neq([Id;2]),
-    "->" = Implies([Id; 2]),
-    "!" = Not(Id),
-    "-" = Neg(Id),
-    "&&" = And([Id;2]),
-    "||" = Or([Id;2]),
-    "^" = Xor([Id;2]),
-    "+" = Add([Id; 2]),
-    "-" = Sub([Id; 2]),
-    "*" = Mul([Id; 2]),
-    "/" = Div([Id; 2]),
-    "%" = Mod([Id; 2]),
-    "min" = Min([Id; 2]),
-    "max" = Max([Id; 2]),
-    "select" = Select([Id; 3]),
-    Var(Symbol),
-  }
-}
-
-impl SynthLanguage for Pred {
-    type Constant = Constant;
-
-    fn constant_to_bool(c: &Self::Constant) -> Option<bool> {
-        Some(c != &0)
-    }
-
-    fn treat_as_pvec(&self) -> bool {
-        match self {
-            Pred::Lt(_) | Pred::Leq(_) | Pred::Eq(_) | Pred::Neq(_) => true,
-            _ => false,
-        }
-    }
-
-    fn eval<'a, F>(&'a self, cvec_len: usize, mut get_cvec: F) -> CVec<Self>
-    where
-        F: FnMut(&'a Id) -> &'a CVec<Self>,
-    {
-        let one = 1.to_i64().unwrap();
-        let zero = 0.to_i64().unwrap();
-        match self {
-            Pred::Lit(c) => vec![Some(*c); cvec_len],
-            Pred::Abs(x) => map!(get_cvec, x => Some(x.abs())),
-            Pred::Lt([x, y]) => {
-                map!(get_cvec, x, y => if x < y {Some(one)} else {Some(zero)})
-            }
-            Pred::Leq([x, y]) => {
-                map!(get_cvec, x, y => if x <= y {Some(one)} else {Some(zero)})
-            }
-            Pred::Eq([x, y]) => {
-                map!(get_cvec, x, y => if x == y {Some(one)} else {Some(zero)})
-            }
-            Pred::Neq([x, y]) => {
-                map!(get_cvec, x, y => if x != y {Some(one)} else {Some(zero)})
-            }
-            Pred::Implies([x, y]) => {
-                map!(get_cvec, x, y => {
-                  let xbool = *x != zero;
-                  let ybool = *y != zero;
-                  if !xbool || ybool {Some(one)} else {Some(zero)}
-                })
-            }
-            Pred::Not(x) => {
-                map!(get_cvec, x => if *x == zero { Some(one)} else {Some(zero)})
-            }
-            Pred::Neg(x) => map!(get_cvec, x => Some(-x)),
-            Pred::And([x, y]) => {
-                map!(get_cvec, x, y => {
-                    let xbool = *x != zero;
-                    let ybool = *y != zero;
-                    if xbool && ybool { Some(one) } else { Some(zero) }
-                })
-            }
-            Pred::Or([x, y]) => {
-                map!(get_cvec, x, y => {
-                    let xbool = *x != zero;
-                    let ybool = *y != zero;
-                    if xbool || ybool { Some(one) } else { Some(zero) }
-                })
-            }
-            Pred::Xor([x, y]) => {
-                map!(get_cvec, x, y => {
-                    let xbool = *x != zero;
-                    let ybool = *y != zero;
-                    if xbool ^ ybool { Some(one) } else { Some(zero) }
-                })
-            }
-            Pred::Add([x, y]) => map!(get_cvec, x, y => x.checked_add(*y)),
-            Pred::Sub([x, y]) => map!(get_cvec, x, y => x.checked_sub(*y)),
-            Pred::Mul([x, y]) => map!(get_cvec, x, y => x.checked_mul(*y)),
-            Pred::Div([x, y]) => map!(get_cvec, x, y => {
-              if y.is_zero() {
-                Some(zero)
-              } else {
-                x.checked_div(*y)
-              }
-            }),
-            Pred::Mod([x, y]) => map!(get_cvec, x, y => {
-              if y.is_zero() {
-                Some(zero)
-              } else {
-                x.checked_rem(*y)
-              }
-            }),
-            Pred::Min([x, y]) => map!(get_cvec, x, y => Some(*x.min(y))),
-            Pred::Max([x, y]) => map!(get_cvec, x, y => Some(*x.max(y))),
-            Pred::Select([x, y, z]) => map!(get_cvec, x, y, z => {
-              let xbool = *x != zero;
-              if xbool {Some(*y)} else {Some(*z)}
-            }),
-            Pred::Var(_) => vec![],
-        }
-    }
-
-    fn initialize_vars(egraph: &mut EGraph<Self, SynthAnalysis>, vars: &[String]) {
-        let consts = vec![
-            Some((-10).to_i64().unwrap()),
-            Some((-1).to_i64().unwrap()),
-            Some(0.to_i64().unwrap()),
-            Some(1.to_i64().unwrap()),
-            Some(2.to_i64().unwrap()),
-            Some(5.to_i64().unwrap()),
-            Some(100.to_i64().unwrap()),
-        ];
-        let cvecs = self_product(&consts, vars.len());
-
-        egraph.analysis.cvec_len = cvecs[0].len();
-
-        for (i, v) in vars.iter().enumerate() {
-            let id = egraph.add(Pred::Var(Symbol::from(v.clone())));
-            let cvec = cvecs[i].clone();
-            egraph[id].data.cvec = cvec;
-        }
-    }
-
-    fn to_var(&self) -> Option<Symbol> {
-        if let Pred::Var(sym) = self {
-            Some(*sym)
-        } else {
-            None
-        }
-    }
-
-    fn mk_var(sym: Symbol) -> Self {
-        Pred::Var(sym)
-    }
-
-    fn is_constant(&self) -> bool {
-        matches!(self, Pred::Lit(_))
-    }
-
-    fn mk_constant(c: Self::Constant, _egraph: &mut EGraph<Self, SynthAnalysis>) -> Self {
-        Pred::Lit(c)
-    }
-
-    fn condition_implies(lhs: &Pattern<Self>, rhs: &Pattern<Self>) -> bool {
-        let mut cfg = z3::Config::new();
-        cfg.set_timeout_msec(1000);
-        let ctx = z3::Context::new(&cfg);
-        let solver = z3::Solver::new(&ctx);
-        let zero = z3::ast::Int::from_i64(&ctx, 0);
-
-        // given that the lhs is true, can we make the rhs false?
-
-        let lhs = egg_to_z3(&ctx, Self::instantiate(lhs).as_ref())
-            ._eq(&zero)
-            .not();
-
-        let rhs = egg_to_z3(&ctx, Self::instantiate(rhs).as_ref())
-            ._eq(&zero)
-            .not();
-
-        let assertion = &lhs;
-
-        solver.assert(assertion);
-
-        if matches!(solver.check(), z3::SatResult::Unsat) {
-            // don't want something that is always false
-            return false;
-        }
-
-        solver.reset();
-        let assertion = &rhs;
-
-        solver.assert(&assertion.not());
-
-        if matches!(solver.check(), z3::SatResult::Unsat) {
-            // don't want something that is always true
-            return false;
-        }
-
-        solver.reset();
-
-        let assertion = &z3::ast::Bool::implies(&lhs, &rhs).not();
-
-        solver.assert(assertion);
-
-        let res = solver.check();
-
-        if matches!(res, z3::SatResult::Unsat) {
-            true
-        } else {
-            false
-        }
-    }
-
-    fn validate(lhs: &Pattern<Self>, rhs: &Pattern<Self>) -> ValidationResult {
-        let mut cfg = z3::Config::new();
-        cfg.set_timeout_msec(1000);
-        let ctx = z3::Context::new(&cfg);
-        let solver = z3::Solver::new(&ctx);
-        let lexpr = egg_to_z3(&ctx, Self::instantiate(lhs).as_ref());
-        let rexpr = egg_to_z3(&ctx, Self::instantiate(rhs).as_ref());
-        solver.assert(&lexpr._eq(&rexpr).not());
-        match solver.check() {
-            z3::SatResult::Unsat => ValidationResult::Valid,
-            z3::SatResult::Unknown => ValidationResult::Unknown,
-            z3::SatResult::Sat => ValidationResult::Invalid,
-        }
-    }
-
-    fn validate_with_cond(
-        lhs: &Pattern<Self>,
-        rhs: &Pattern<Self>,
-        cond: &Pattern<Self>,
-    ) -> ValidationResult {
-        assert!(cond.to_string().len() > 2, "Conditional pattern: {}", cond);
-        let mut cfg = z3::Config::new();
-        cfg.set_timeout_msec(1000);
-        let ctx = z3::Context::new(&cfg);
-        let solver = z3::Solver::new(&ctx);
-        let zero = z3::ast::Int::from_i64(&ctx, 0);
-        let cexpr =
-            z3::ast::Bool::not(&egg_to_z3(&ctx, Self::instantiate(cond).as_ref())._eq(&zero));
-
-        let lexpr = egg_to_z3(&ctx, Self::instantiate(lhs).as_ref());
-        let rexpr = egg_to_z3(&ctx, Self::instantiate(rhs).as_ref());
-        solver.assert(&z3::ast::Bool::implies(&cexpr, &lexpr._eq(&rexpr)).not());
-
-        let res = solver.check();
-
-        match res {
-            z3::SatResult::Unsat => ValidationResult::Valid,
-            z3::SatResult::Unknown => ValidationResult::Unknown,
-            z3::SatResult::Sat => ValidationResult::Invalid,
-        }
-    }
-}
-
-fn egg_to_z3<'a>(ctx: &'a z3::Context, expr: &[Pred]) -> z3::ast::Int<'a> {
-    let mut buf: Vec<z3::ast::Int> = vec![];
-    let zero = z3::ast::Int::from_i64(ctx, 0);
-    let one = z3::ast::Int::from_i64(ctx, 1);
-    for node in expr.as_ref().iter() {
-        match node {
-            Pred::Lit(c) => buf.push(z3::ast::Int::from_i64(ctx, c.to_i64().unwrap())),
-            Pred::Abs(x) => {
-                let l = &buf[usize::from(*x)];
-                buf.push(z3::ast::Bool::ite(
-                    &z3::ast::Int::lt(l, &zero),
-                    &z3::ast::Int::unary_minus(l),
-                    l,
-                ))
-            }
-            Pred::Lt([x, y]) => {
-                let l = &buf[usize::from(*x)];
-                let r = &buf[usize::from(*y)];
-                buf.push(z3::ast::Bool::ite(&z3::ast::Int::lt(l, r), &one, &zero))
-            }
-            Pred::Leq([x, y]) => {
-                let l = &buf[usize::from(*x)];
-                let r = &buf[usize::from(*y)];
-                buf.push(z3::ast::Bool::ite(&z3::ast::Int::le(l, r), &one, &zero))
-            }
-            Pred::Eq([x, y]) => {
-                let l = &buf[usize::from(*x)];
-                let r = &buf[usize::from(*y)];
-                buf.push(z3::ast::Bool::ite(&z3::ast::Int::_eq(l, r), &one, &zero))
-            }
-            Pred::Neq([x, y]) => {
-                let l = &buf[usize::from(*x)];
-                let r = &buf[usize::from(*y)];
-                buf.push(z3::ast::Bool::ite(&z3::ast::Int::_eq(l, r), &zero, &one))
-            }
-            Pred::Implies([x, y]) => {
-                let l = &buf[usize::from(*x)];
-                let r = &buf[usize::from(*y)];
-                let l_not_z = z3::ast::Bool::not(&l._eq(&zero));
-                let r_not_z = z3::ast::Bool::not(&r._eq(&zero));
-                buf.push(z3::ast::Bool::ite(
-                    &z3::ast::Bool::implies(&l_not_z, &r_not_z),
-                    &one,
-                    &zero,
-                ))
-            }
-            Pred::Not(x) => {
-                let l = &buf[usize::from(*x)];
-                buf.push(z3::ast::Bool::ite(&l._eq(&zero), &one, &zero))
-            }
-            Pred::Neg(x) => buf.push(z3::ast::Int::unary_minus(&buf[usize::from(*x)])),
-            Pred::And([x, y]) => {
-                let l = &buf[usize::from(*x)];
-                let r = &buf[usize::from(*y)];
-                let l_not_z = z3::ast::Bool::not(&l._eq(&zero));
-                let r_not_z = z3::ast::Bool::not(&r._eq(&zero));
-                buf.push(z3::ast::Bool::ite(
-                    &z3::ast::Bool::and(ctx, &[&l_not_z, &r_not_z]),
-                    &one,
-                    &zero,
-                ))
-            }
-            Pred::Or([x, y]) => {
-                let l = &buf[usize::from(*x)];
-                let r = &buf[usize::from(*y)];
-                let l_not_z = z3::ast::Bool::not(&l._eq(&zero));
-                let r_not_z = z3::ast::Bool::not(&r._eq(&zero));
-                buf.push(z3::ast::Bool::ite(
-                    &z3::ast::Bool::or(ctx, &[&l_not_z, &r_not_z]),
-                    &one,
-                    &zero,
-                ))
-            }
-            Pred::Xor([x, y]) => {
-                let l = &buf[usize::from(*x)];
-                let r = &buf[usize::from(*y)];
-                let l_not_z = z3::ast::Bool::not(&l._eq(&zero));
-                let r_not_z = z3::ast::Bool::not(&r._eq(&zero));
-                buf.push(z3::ast::Bool::ite(
-                    &z3::ast::Bool::xor(&l_not_z, &r_not_z),
-                    &one,
-                    &zero,
-                ))
-            }
-            Pred::Add([x, y]) => buf.push(z3::ast::Int::add(
-                ctx,
-                &[&buf[usize::from(*x)], &buf[usize::from(*y)]],
-            )),
-            Pred::Sub([x, y]) => buf.push(z3::ast::Int::sub(
-                ctx,
-                &[&buf[usize::from(*x)], &buf[usize::from(*y)]],
-            )),
-            Pred::Mul([x, y]) => buf.push(z3::ast::Int::mul(
-                ctx,
-                &[&buf[usize::from(*x)], &buf[usize::from(*y)]],
-            )),
-            Pred::Div([x, y]) => {
-                let l = &buf[usize::from(*x)];
-                let r = &buf[usize::from(*y)];
-                buf.push(z3::ast::Bool::ite(
-                    &r._eq(&zero),
-                    &zero,
-                    &z3::ast::Int::div(l, r),
-                ))
-            }
-            Pred::Mod([x, y]) => {
-                let l = &buf[usize::from(*x)];
-                let r = &buf[usize::from(*y)];
-                buf.push(z3::ast::Bool::ite(
-                    &r._eq(&zero),
-                    &zero,
-                    &z3::ast::Int::modulo(l, r),
-                ))
-            }
-            Pred::Min([x, y]) => {
-                let l = &buf[usize::from(*x)];
-                let r = &buf[usize::from(*y)];
-                buf.push(z3::ast::Bool::ite(&z3::ast::Int::le(l, r), l, r))
-            }
-            Pred::Max([x, y]) => {
-                let l = &buf[usize::from(*x)];
-                let r = &buf[usize::from(*y)];
-                buf.push(z3::ast::Bool::ite(&z3::ast::Int::le(l, r), r, l))
-            }
-            Pred::Select([x, y, z]) => {
-                let cond = z3::ast::Bool::not(&buf[usize::from(*x)]._eq(&zero));
-                buf.push(z3::ast::Bool::ite(
-                    &cond,
-                    &buf[usize::from(*y)],
-                    &buf[usize::from(*z)],
-                ))
-            }
-            Pred::Var(v) => buf.push(z3::ast::Int::new_const(ctx, v.to_string())),
-        }
-    }
-    buf.pop().unwrap()
-}
-
 #[cfg(test)]
 #[path = "./recipes/halide.rs"]
 mod halide;
@@ -404,9 +5,10 @@ mod halide;
 #[allow(unused_imports)]
 mod test {
     use crate::halide::{
-        halide_rules, halide_rules_for_caviar_conditional, halide_rules_for_caviar_total_only,
+        halide_limited_grammar, halide_rules, halide_rules_for_caviar_conditional,
+        halide_rules_for_caviar_total_only,
     };
-    use crate::{egg_to_z3, Pred};
+    use ruler::halide::{egg_to_z3, Pred};
     use std::{
         sync::Arc,
         time::{Duration, Instant},
